refactor(mcp): rename the Reach tab to Pieces
"Reach" reads as a verb before it reads as a noun, and names no object.
The tab lists pieces, so it is called Pieces — the word every user already
owns. The verb survives in the body copy, matching the Connect and
Connections tabs.

The route moves with the label (/mcp-server/pieces), and the row model
keeps ReachablePiece: it describes the semantics, not the tab.
