+++ packages/web/public/locales/en/translation.json
@@ -2634,7 +2634,6 @@
   "revokeSelectedCount": "{count, plural, =1 {Revoke 1} other {Revoke #}}",
   "revokedGrants": "{count, plural, =1 {1 connection} other {# connections}}",
   "{name} · you": "{name} · you",
-  "Reach": "Reach",
   "Every piece a connected client can reach, and every action inside it.": "Every piece a connected client can reach, and every action inside it.",
   "This page is a mirror — a platform admin decides what is on the list.": "This page is a mirror — a platform admin decides what is on the list.",
   "Nothing below can run right now": "Nothing below can run right now",
@@ -2647,8 +2646,8 @@
   "Every piece below is reachable by any connected client. Restricting the list to a chosen set is an enterprise feature.": "Every piece below is reachable by any connected client. Restricting the list to a chosen set is an enterprise feature.",
   "This project's pieces are controlled by a Piece Set.": "This project's pieces are controlled by a Piece Set.",
   "Review piece set": "Review piece set",
-  "reachDestructiveActionCount": "{count, plural, =1 {1 destructive} other {# destructive}}",
-  "reachActionCount": "{count, plural, =1 {1 action} other {# actions}}",
+  "pieceDestructiveActionCount": "{count, plural, =1 {1 destructive} other {# destructive}}",
+  "pieceActionCount": "{count, plural, =1 {1 action} other {# actions}}",
   "Opens Cursor and writes the server into ~/.cursor/mcp.json.": "Opens Cursor and writes the server into ~/.cursor/mcp.json.",
   "Paste the server URL as a custom connector. This client dials your server from the internet, so localhost will not reach it.": "Paste the server URL as a custom connector. This client dials your server from the internet, so localhost will not reach it.",
   "Paste the server URL as a connector. ChatGPT dials your server from the internet, so localhost will not reach it.": "Paste the server URL as a connector. ChatGPT dials your server from the internet, so localhost will not reach it.",
