+++ packages/web/src/app/routes/mcp-server/index.tsx
@@ -9,7 +9,7 @@ import { GrantsTab } from './grants/grants-tab';
 import { useMcpNav } from './mcp-nav';
 import { useMcpServerUrl } from './mcp-server-url';
 import { PageBand } from './page-band';
-import { ReachTab } from './reach/reach-tab';
+import { PiecesTab } from './pieces/pieces-tab';
 
 export default function McpServerPage() {
   const { serverUrl, isReachableFromInternet } = useMcpServerUrl();
@@ -26,8 +26,8 @@ export default function McpServerPage() {
               <TabsTrigger variant="outline" value="connect">
                 {t('Connect')}
               </TabsTrigger>
-              <TabsTrigger variant="outline" value="reach">
-                {t('Reach')}
+              <TabsTrigger variant="outline" value="pieces">
+                {t('Pieces')}
               </TabsTrigger>
               <TabsTrigger variant="outline" value="connections">
                 {t('Connections')}
@@ -37,8 +37,8 @@ export default function McpServerPage() {
         </PageBand>
       </div>
       <div className="w-full">
-        {nav.tab === 'reach' ? (
-          <ReachTab
+        {nav.tab === 'pieces' ? (
+          <PiecesTab
             projectId={nav.projectId}
             onSelectProject={nav.selectProject}
           />
