+++ packages/web/src/app/routes/mcp-server/mcp-nav.ts
@@ -3,7 +3,7 @@ import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
 import { authenticationSession } from '@/lib/authentication-session';
 
 function toTab(value: string | undefined): McpTab {
-  return value === 'connections' || value === 'reach' ? value : 'connect';
+  return value === 'connections' || value === 'pieces' ? value : 'connect';
 }
 
 export function useMcpNav(): McpNav {
@@ -25,7 +25,7 @@ export function useMcpNav(): McpNav {
   };
 }
 
-export type McpTab = 'connect' | 'reach' | 'connections';
+export type McpTab = 'connect' | 'pieces' | 'connections';
 
 export type McpView = 'landing' | 'browse' | 'client';
