+++ packages/web/src/app/routes/mcp-server/pieces/project-picker.tsx
